requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
python-dotenv==1.0.0
pymongo==4.5.0
PyJWT==2.8.0
passlib==1.7.4
bcrypt==4.0.1
google-genai==1.0.0
gunicorn==21.2.0
bson==0.5.10
requests==2.31.0
fpdf==1.7.2
